Added the von_mises distribution for 2-d and 3-d
SVN-Revision: 26968

+++ contrib/brl/bbas/bsta/Templates/bsta_gaussian_sphere+double.2-.cxx
@@ -3,4 +3,4 @@
 // Approved for public release, distribution unlimited (DISTAR Case 14389)
 //
 
-BSTA_GAUSSIAN_SPHERE_INSTANTIATE(float, 2);
+BSTA_GAUSSIAN_SPHERE_INSTANTIATE(double, 2);
